Move eacs_block_update to its own file
