+++ src/libvgmstream.vcproj
@@ -1591,6 +1591,10 @@
                     >
                 </File>
                 <File
+                    RelativePath=".\layout\blocked_ea_1snh.c"
+                    >
+                </File>
+                <File
                     RelativePath=".\layout\blocked_vgs.c"
                     >
                 </File>
