+++ src/libvgmstream.vcxproj
@@ -471,6 +471,7 @@
     <ClCompile Include="layout\bdsp_blocked.c" />
     <ClCompile Include="layout\blocked.c" />
     <ClCompile Include="layout\blocked_awc.c" />
+    <ClCompile Include="layout\blocked_ea_1snh.c" />
     <ClCompile Include="layout\blocked_vgs.c" />
     <ClCompile Include="layout\caf_blocked.c" />
     <ClCompile Include="layout\de2_blocked.c" />
