+++ src/libvgmstream.vcxproj.filters
@@ -925,6 +925,9 @@
     <ClCompile Include="layout\blocked_awc.c">
       <Filter>layout\Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="layout\blocked_ea_1snh.c">
+      <Filter>layout\Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="layout\blocked_vgs.c">
       <Filter>layout\Source Files</Filter>
     </ClCompile>
